chore: Move Breadcrumbs component to its own folder
