+++ src/theme/BlogLayout/index.js
@@ -2,14 +2,12 @@ import React from 'react';
 import clsx from 'clsx';
 import Layout from '@theme/Layout';
 import BlogSidebar from '@theme/BlogSidebar';
-import BlogBreadcrumbs from '../BlogBreadcrumbs';
 
 export default function BlogLayout(props) {
   const {sidebar, toc, children, ...layoutProps} = props;
   const hasSidebar = sidebar && sidebar.items.length > 0;
   return (
     <Layout {...layoutProps}>
-      <BlogBreadcrumbs />
       <div className="container margin-vert--lg">
         <div className="row">
           <BlogSidebar sidebar={sidebar} />
